requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pandas>=1.5.0
requests>=2.28.0
scikit-learn>=1.1.0
xgboost>=1.6.0
joblib>=1.1.0
python-dotenv>=1.0.0
numpy>=1.22.0
pymongo>=4.0.0
APScheduler>=3.9.0
pyarrow>=10.0.0 
matplotlib>=3.5.0 